+++ docs/source/api/DomainMind.rst
@@ -1,0 +1,93 @@
+DomainMind package
+==================
+
+Submodules
+----------
+
+DomainMind.DataLoad module
+--------------------------
+
+.. automodule:: DomainMind.DataLoad
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+DomainMind.Interface module
+---------------------------
+
+.. automodule:: DomainMind.Interface
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+DomainMind.LLMConfig module
+---------------------------
+
+.. automodule:: DomainMind.LLMConfig
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+DomainMind.MultiQueryGenChain module
+------------------------------------
+
+.. automodule:: DomainMind.MultiQueryGenChain
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+DomainMind.MultiQueryRetrievalChain module
+------------------------------------------
+
+.. automodule:: DomainMind.MultiQueryRetrievalChain
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+DomainMind.OutputProcess module
+-------------------------------
+
+.. automodule:: DomainMind.OutputProcess
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+DomainMind.RAGChain module
+--------------------------
+
+.. automodule:: DomainMind.RAGChain
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+DomainMind.RetrieverConfig module
+---------------------------------
+
+.. automodule:: DomainMind.RetrieverConfig
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+DomainMind.SummaryChain module
+------------------------------
+
+.. automodule:: DomainMind.SummaryChain
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+DomainMind.VectorDB module
+--------------------------
+
+.. automodule:: DomainMind.VectorDB
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: DomainMind
+   :members:
+   :show-inheritance:
+   :undoc-members:
