+++ docs/source/api/create_vectordb.rst
@@ -1,0 +1,7 @@
+create\_vectordb module
+=======================
+
+.. automodule:: create_vectordb
+   :members:
+   :show-inheritance:
+   :undoc-members:
