+++ docs/source/api/main.rst
@@ -1,0 +1,7 @@
+main module
+===========
+
+.. automodule:: main
+   :members:
+   :show-inheritance:
+   :undoc-members:
