--- docs/source/api/src.rst
@@ -1,93 +1,0 @@
-src package
-===========
-
-Submodules
-----------
-
-src.DataLoad module
--------------------
-
-.. automodule:: src.DataLoad
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-src.Interface module
---------------------
-
-.. automodule:: src.Interface
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-src.LLMConfig module
---------------------
-
-.. automodule:: src.LLMConfig
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-src.MultiQueryGenChain module
------------------------------
-
-.. automodule:: src.MultiQueryGenChain
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-src.MultiQueryRetrievalChain module
------------------------------------
-
-.. automodule:: src.MultiQueryRetrievalChain
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-src.OutputProcess module
-------------------------
-
-.. automodule:: src.OutputProcess
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-src.RAGChain module
--------------------
-
-.. automodule:: src.RAGChain
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-src.RetrieverConfig module
---------------------------
-
-.. automodule:: src.RetrieverConfig
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-src.SummaryChain module
------------------------
-
-.. automodule:: src.SummaryChain
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-src.VectorDB module
--------------------
-
-.. automodule:: src.VectorDB
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: src
-   :members:
-   :show-inheritance:
-   :undoc-members:
