+++ docs/source/conf.py
@@ -9,16 +9,22 @@
 import os
 import sys
 
+from click import prompt
+
 # sys.path.insert(0, os.path.abspath("../.."))
 # sys.path.insert(0, os.path.abspath("../../src"))
 
 # Always resolve paths relative to this conf.py file
 conf_dir = os.path.dirname(os.path.abspath(__file__))
 project_root = os.path.abspath(os.path.join(conf_dir, "../.."))
-src_dir = os.path.abspath(os.path.join(conf_dir, "../../src"))
+src_dir = os.path.abspath(os.path.join(conf_dir, "../../DomainMind"))
+data_dir = os.path.abspath(os.path.join(conf_dir, "../../data"))
+prompts_dir = os.path.abspath(os.path.join(conf_dir, "../../prompts"))
 
 sys.path.insert(0, project_root)
 sys.path.insert(0, src_dir)
+sys.path.insert(0, data_dir)
+sys.path.insert(0, prompts_dir)
 
 
 project = "DomainMind"
