--- src/DomainMind/LLMConfig.py
@@ -1,3 +1,0 @@
-from langchain_community.chat_models import ChatOllama
-
-llm = ChatOllama(model="qwen3:8b", base_url="http://localhost:11434", streaming=True)
